Move text-emphasis-style variables out of mako templates
