+++ components/style/values/computed/text.rs
@@ -15,7 +15,7 @@ use values::generics::text::InitialLetter as GenericInitialLetter;
 use values::generics::text::LineHeight as GenericLineHeight;
 use values::generics::text::MozTabSize as GenericMozTabSize;
 use values::generics::text::Spacing;
-use values::specified::text::{TextOverflowSide, TextDecorationLine};
+use values::specified::text::{TextOverflowSide, TextDecorationLine, FillMode, ShapeKeyword};
 
 pub use values::specified::TextAlignKeyword as TextAlign;
 
@@ -151,3 +151,16 @@ impl TextDecorationsInEffect {
 
 /// A specified value for the `-moz-tab-size` property.
 pub type MozTabSize = GenericMozTabSize<NonNegativeNumber, NonNegativeLength>;
+
+#[derive(Clone, Debug, MallocSizeOf, PartialEq, ToCss)]
+pub enum TextEmphasisStyle {
+    Keyword(KeywordValue),
+    None,
+    String(String),
+}
+
+#[derive(Clone, Debug, MallocSizeOf, PartialEq, ToCss)]
+pub struct KeywordValue {
+    pub fill: FillMode,
+    pub shape: ShapeKeyword,
+}
